requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dotenv==0.9.9
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.31.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.4
